% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterMosaicFromFiles.R
\name{mosaicFromFiles}
\alias{mosaicFromFiles}
\title{Raster mosaic from file names}
\usage{
mosaicFromFiles(x, cropOut = NULL, fun = "mean", tolerance = 0.05,
  filename = paste(tempfile(), ".grd", sep = ""),
  overwrite = file.exists(filename), ...)
}
\arguments{
\item{x}{a string specifying raster files}

\item{cropOut}{area to crop the output raster to}

\item{fun}{Function, see \code{\link[raster]{mosaic}}}

\item{tolerance, }{numeric, see \code{\link[raster]{mosaic}}}

\item{filename}{Output file, see \code{\link[raster]{writeRaster}}}

\item{overwrite}{Logical, see \code{\link[raster]{writeRaster}}}

\item{...}{Additional arguments for \code{\link[raster]{writeRaster}}}
}
\value{
A \code{\link[raster]{raster}}
}
\description{
Mosaic raster files into a new raster with a larger spatial extent
}
\seealso{
\code{\link[raster]{mosaic}}, \code{\link[raster]{writeRaster}}
}
